refactor: Simplify relayer partial fill handling
Checking for existing fills in the pending transaction queue is only
necessary because of some test cases that instrument the relayer in
specific ways that results in the relayer enqueueing multiple fills
for the same deposit. This is now how the relayer works in production.

Flushing the transaction queue at the start of each run allows for the
relayer to assume that it has no previously-enqueued fills.