feat: :sparkles: programmatically track provider errors in optional cache
This change is meant to lay the skeleton for dynamically determining which order to rank providers for a given chain. This first PR is aimed at adding a key in the cache when errors occur. An exponential decay is used to store a count of the errors incurred by a given provider