fix: Use RL relayer address for basic gas cost estimation (#1036)
It's invalid to use the current relayer's EOA as the simulated relayer
address because that address may not hold USDC on all chains. Instead,
default to the main RL address, which is known to have all relevant
tokens on all chains.

The relayer's own address will be used for dynamic message-based
simulation.