improve(ProviderUtils): Temporarily ignore block size field
Alchemy have been disagreeing with other providers on Arbitrum for a few
days now; it causes higher RPC usage and unnecessary log noise. Disable
it for now, with the option of reverting in future (if they fix it) or
retaining it indefinitely, since we don't actually rely on this field.