chore: Drop Across API timeout from 60s to 3s (#1058)
Most API queries should resolve in under 1 second, but there are
occasions where we log that we hit the 60s timeout. This is a major
slowdown for the bots. Since we now rotate to the next available route,
if the first query doesn't resolve shortly after it was sent, we should
just fail quickly and move on instead.