refactor(Dataworker): Replace ProfitClient with PriceClient (#1073)
The Dataworker only relies indirectly on the ProfitClient for its
ability to query token prices. This imposes some test/mock burden on the
ProfitClient. This change hoists the ProfitClient out of the Dataworker
and instead subs in the underlying PriceClient, which the ProfitClient
also uses directly.

As a side benefit, the proposer is currently the only component
dependent on token prices, so token price lookups can be relegated to
the code that actually needs it.