fix(ProfitClient): Support non-unique token addresses (#1076)
In cases like ETH/WETH, the underlying mainnet contract address used to
resolve the token price is the same. Therefore, mapping from contract
address to L1Token doesn't work very well for this purpose. Invert the
token mapping so that the token symbol is unique, allowing for multiple
symbols to map to a single address. This also helps in the case of test
tokens that are temporarily mapped to an existing token.