fix(zksync-finalizer): Identify correct log index of ZkSync withdrawal when there are multiple withdrawals in same L2 transaction
The current code doesn't handle the case where there are multiple `TokensBridged` events within a single ZkSync transaction. This is because it always returns the `withdrawalIdx = 0` for all batched `TokensBridged`.

I believe this is because the `.reverse()` always causes the subsequent `.find()` to return index 0.

This PR simplifies the logic a lot and simply:
- Find all `L1MessageSentEvents` preceding the `TokensBridged` event.
- Return `withdrawalIdx` equal to the # of above preceding `L1MessageSentEvents`. This should return a unique withdrawal index per TokensBridged