improve(relayer): Don't capture token shorfall when zero-filling
I think this can address the failure case where we submit a zero fill and send an inventory rebalance to an L2 in the same Relayer run when we're short balance there.

This PR would lead to fewer slow fills initiated that end up getting fast-filled, but would lead to better capital efficiency for the relayer.