improve(BlockFinder): Permit supplying block range hints
The BlockFinder has historically had to try to estimate the lower bound
of its search based on average block time estimates. On occasion these
estimates can be inaccurate and can result in significantly overshooting
the lower bound, leading to an extended search. This is mostly
observable on chains where the average block time is inconsistent (i.e.
testnets) or has changed notably over time (Arbitrum or Optimism).

This change allows the caller to supply a hint to the BlockFinder about
the probable upper and/or lower bounds, such that any error in its own
estimation is minimised.