fix: Supply cache instance to getCachedBlockForTimestamp() (#1089)
Without this we don't cache the timestamp -> blocknumber mappings, so we rely on the underlying RPC caching to retain the responses to the eth_getBlock queries that we issue. The are two reasons this is less desirable:

- Cached RPC queries are typically ejected much more regularly (default 1 hour, vs 1 week for cached block timestamps).
- Timestamp to block number resolution usually takes a different search path on each query, so there is a high chance of RPC cache misses even when trying to resolve the same timestamp.

This is one of the underlying reasons for the elevated RPC consumption of late.