improve(BlockFinder): Permit supplying block range hints (#1090)
The BlockFinder has historically tried to find a sensible lower bound
for its search space based on average block times. On occasion these
estimates can be inaccurate and can result in significantly overshooting
the lower bound, leading to an intensive and slow search. This is mostly
observable on chains where the average block time is inconsistent (i.e.
testnets) or has changed notably over time (Arbitrum or Optimism).

This change allows the caller to supply a hint to the BlockFinder about
the probable upper and/or lower bounds of the requested timestamp. Any
supplied hints are optimistically pre-cached in the hope that they might
tightly bound the requested timestamp. If the hinted lower bound
successfully bounds the timestamp then the estimation process will be
skipped.