feat: Support instantiating void signers
tldr there are various scripts that don't need a real signer, so don't
require a private key or mnemonic to be loaded in these cases. This
makes life easier for people who just want to pull down the repo and
run a script (i.e. validateRootBundle.ts).

The ethers Wallet type inherits from Signer, but it imposes additional
requirements like having a private key. By deferring to the Signer type,
we also permit void signers. This is nice for scenarios where a private
key is not needed, which is in fact the vast majority of the usage in
the bot.