fix(Dataworker): Set SpokePool end blocks more conservatively (#1108)
This change ensures that Dataworker proposal block range identification
is set realistically relative to the event search config for each
SpokePoolClient.

HubPoolClient and SpokePoolClient instantiation was recently updated
such that their search configs are now synchronised, and determined
close to startup. Previously, the SpokePoolClient's search config would
be determined much later because it had to wait for the initial
HubPoolClient update to complete, and that can span several mainnet
blocks. This meant that there was typically a 0 delta between the
toBlock search config for each SpokePoolClient and its recorded
latestBlockNumber. After the recent update, these values are now
divergent.

This divergence can result in a situation where each SpokePoolClient may
not have sufficient data to correctly match each fill with a deposit,
and this has resulted in several disputes over the past few days.

Fixes ACX-1755