feat: Support instantiating void signers (#1106)
The ethers Wallet type inherits from Signer, but it imposes additional
requirements like having a private key. By deferring to the Signer type,
we also permit void signers. This is nice for scenarios where a private
key is not needed, which is in fact the vast majority of the usage in
the bot.