fix: Use disputer address in validateRootBundle script (#1112)
The validateRootBundle script inherits a lot of core functionality,
including the TokenClient. The TokenClient automatically seeks to put
the necessary token approvals in place for handling the ABT token, but
that's not strictly required for the validateRootBundle script since it
never actually makes any transactions. Work around this by specifying
the proposer address when instantiating the VoidSigner instance.