refactor: Reduce use of disposable BigNumber instances (#1105)
ethers provides predefined BNs for 0 and 1, and we import these via the
Across SDK. Roll these out as far and wide as possible, so as to reduce
the unnecessary instantiation of new short-lived BNs. This will admittedly
only have a small impact on memory and CPU consumption, but is still
worth doing (especially in cases where the BNs were previously instantiated
in a tight loop).