fix(InventoryClient): Don't try to query l2 token for l1 token on chain that it isn't enabled for
If `constants-v2` and `sdk-v2` aren't aware of a token's address on an L2 (example, WBTC on Base/8453), then attempting to call `HubPoolClient.getDestinationTokenForL1Token()` will throw an error. We should handle this early and exit early like we do in other functions.

The offending function fixed in this PR is `getChainDistribution()`, which is called by `getTokenDistributionPerL1Token()`