chore: Update default signer to secret (#1057)
This changes the default method of private key derivation from
`mnemonic` to `secret`. Secret is preferred because it allows the user
to store the private key separately from their .env file, which might
otherwise be edited frequently or exposed via screen sharing.

While here, loosen the regex for the private key so that the leading 0x
becomes optional. Anything that doesn't match the private key regex is
assumed to be a mnemonic.