fix: Increase Polygon lookback (#1150)
The Polygon finalizer instance limits its lookback to head - lookback,
which is currently set to 1 day. If the finalizer is down or ineffective
for > 1 day then this risks ignoring transactions that would otherwise
be able to be finalised. There's very limited benefit to limiting the
lookback to only 1 day, so pad it out to 5.