refactor: Remove relayerFeePct ordering in relayer (#1194)
The comment attached to this ordering says that it's done to prioritise
the most profitable deposits first. This is only a very coarse way of
sorting because it ignores the amount being bridged.

Sorting this early isn't compatible with v3 deposits because their
profitability is determined later in the relayer run. In any case,
ordering deposits by profitability doesn't seem to have much benefit in
practice, so remove it for now.