Re-work contraints & test
Use SpokePool timestamps instead of host timestamps due to issues in
making the tests work properly. This test worked OK in isolation, but in
a parallel environment or when other tests are also run it seems to go
awry because of the increasing gap between host time and chain time.

The downsides of using SpokePool are limited and probably have no real
impact, since the expiries should be so far ahead that we rarely (if
ever) see a scenario where we could have made a fill but end up missing
it.