fix: Restore gas cost reporting
This was subtly broken back in November the ProfitClient was updated to
additionally report on both the estimated gas cost for making a fill.
The name nativeGasCost had its meaning changed slightly, where it went
from reporting on units of gas token to units of gas.

The relevant commit was:
  bd6cb886b4065612fe59f94ea1aae9d1fcd26a61