improve: Add historical deposit sanity check when querying historical fill
This code path is hit when a deposit can't be found to validate a slow fill request or a fill. There is a very very slight chance this could locate a deposit LATER than the bundle end block, in the event someone sent a fill, followed by a deposit that happened to match it after the fact.

This could be used to grief the dataworker by inducing a dispute