feat: Support submitting private transactions
This change introduces the ability to define custom RPC providers to be
used when submitting transactions. Custom providers are slotted under
the PRIVACY namespace, since that's the anticipated use case for them.

This is held fairly generic within ProviderUtils by scoping the change
as a generic namespace, rather than introducing any concept of privacy.