fix(ProfitClient): Correct gas estimation errors
Gas estimation for fills on most chains has been on the low side and has
not been representative of the actual cost of filling. This has been
traced was traced to the fill amount being set to 1 unit of the
destination token, which leads to fee rounding in the SpokePool, in turn
producing a lower than expected gas estimate. This is not representative
of the true cost of a fill.

Having addressed the first issue, the use of WETH on mainnet destined
for a recipient with 0 balance has proven to produce over-estimates for
the fill cost. This seems to be due to the marginal cost of initialising
a recipient's ETH balance to a non-zero value, with the delta being
about 25k gas for doing so (verified by testing against an "unused" EOA
vs. one with an ETH balance. WETH additionally incurs the marginal
withdrawal (unwrap) expense when sending to an EOA, leading to an
eventual estimate of about 140k gas on mainnet. This is a bit on the
high side because _most_ recipients would be unlikely to have a zero ETH
balance, but their ERC20 balances are probably more regularly going to
be zero.

When testing with USDC on mainnet, the resulting estimate is instead
~130k gas. This seems like a more reasonable assumption for the
ProfitClient to make, based on actual onchain sampling of fills.

tldr: the before/after fill cost for mainnet is:
Before: ~95k gas
After: ~130k gas