refactor: Simplify FillUtils getUnfilledDeposits() (#1258)
This is an incremental change to drop some unneeded members from the
return type, and to return an Object of destination chain => unfilled
deposits array. This permits the caller to keep the unfilled deposits
segregated, which fits future use cases pretty well.

This is take #2 at this change, since the original commit incorrectly
handled the deposit filtering based on the block lookback by applying
the wrong "low block" for each chain Id.