improve(relayer): Limit HubPoolClient lookback (#1275)
The HubPoolClient now has logic to override the requested lookback for
specific events if it has not previously been updated. This is the same
logic that is implemented for the SpokePoolClients. As a result, it's
now possible to be much more restrictive on the lookback for the
HubPoolClient. The relayer can benefit from this be only querying
RootBundle events for the past ~12 hours, which should be sufficient for
the InventoryManager to determine repayment chains.