feat: Add optional secondary simulation option in MulticallerClient
Can be used by Relayer to add a second simulation for certain transactions that we expect ethers to not always be correct about. For example, filling deposits with messages sent to Proxy recipients on Base has been succeeding in the ethers txn simulation but failing when actually submitted.

It's unclear whether the recipient proxy or the Base network is at fault but we do know that the Alchemy gas estimator SDK and the ethers gas estimator disagree, with the former catching the failed txn and the latter not.