improve(relayer): Batch query deposit fill status (#1338)
The relayer currently implements late-stage checking of the fill status
on a per-deposit basis. This is fairly slow and inefficient, but does
successfully protect against an avalanche of failed fill simulations
that result from overlapping bot runs.

This change implements batched querying of all deposits for each
destination chain. This is significantly more efficient in terms of RPC
consumption, and should allow for subsequent "already filled" filtering
to occur much faster. It's also a key enabler for some subsequent
efficiency improvements.