improve(TokenClient): Cache allowance and bondToken eth_call data (#1322)
* improve(TokenClient): Cache allowance and bondToken eth_call data

These values should rarely change. Follows strategy implemented in #1282 in the BaseAdapter (inventory manager)

* Update TokenClient.ts

* Update TokenClient.ts

* Update TokenClient.ts

* Revert "Update TokenClient.ts"

This reverts commit d8b0f00554c9aa26a7def4a22ec61aca3f9c71a5.

* Revert "Update TokenClient.ts"

This reverts commit c1703fb0eca15cd255f7cf239d0de2d961f81051.

* Revert "Update TokenClient.ts"

This reverts commit 2b2b325ac83c502cb9ccd3b4f6b6610d191379ed.

* Refactor because using hard-coded blockTag breaks tests

While fixing the block tag to the nearest 100_000th block when querying bondToken is nice, it breaks tests and I don't think they're worth fixing. The main motivation for taking advantage of the eth_call specific-blockTag caching is to make the code cleaner but I think this commit's refactor is nice too

* Update TokenClient.ts

* check for MAX_SAFE_ALLOWANCE

---------

Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>