improve(relayer): Batch-compute all expected LP fees
The introduction of v3 temporarily removed the ability to batch compute
LP fees on deposit ingestion because the (re)payment chain is no longer
determined by the deposit destination chain. Instead, the relayer has
been computing LP fees at least once per candidate fill, and often twice
in the event that the preferred repayment chain is mainnet.

One of the major downsides of computing LP fees individually is that the
BlockFinder must be invoked to resolve the quoteTimestamp to a mainnet
block number. There are significant savings when this is batched due to
duplicate quoteTimestamp consolidation in the HubPoolClient.

This PR restores batch computation of LP fees by aggregating all
outstanding deposits and computing them in bulk. This is performed for
both destination chain and mainnet. This permits the relayer to
subsequently evaluate profitability on the fly, without needing to yield
for LP fee computation. The net effect should thus be that relayer
runtimes are more consistent and less reactive to a surge in the number
of unfilled deposits.