improve(FillUtils): Protect getUnfilledDeposits() against RPC error
If the deposit fill status cannot be queried on the destination chain,
simply assume all deposits are unfilled. This allows execution to
continue, and the complete set of resulting deposits with unknown status
will likely be reduced by subsequent filtering anyway (i.e. based on
supported token, number of deposit confirmations, profitability, ...).
The ultimate fill status for the residual deposits will be inferred at
the time of fill simulation on the destination chain.