chore(Dataworker): Explicitly require SpokePoolClient update
A pending update to the sdk will allow the SpokePoolClient to continue
in an un-updated state. The relayer is OK with this but the Dataworker
requires a complete set of updated SpokePoolClient instances. Pre-empt
this pending change by requiring explicitly that all Dataworker
SpokePoolClient instances updated successfully. This is a case where
it's OK (and expected) to fail hard.