improve(BundleDataClient): Cache Arweave requests (#1390)
Cache the arweave request as a Promise so that we only make 1 request to Arweave per bundle block range. If these requests are sent in parallel, then the cache should be set to a Promise by the first request to send, and then when this request resolves, all parallel requests can use its response.