refactor(relayer): Drop unfilled deposit lookback filtering (#1401)
This function invokes a BlockFinder instance on each chain and proceeds
to resolve the lookback time delta to a block. It can be expensive to do
this on chains where the average block times are not super consistent,
so it sometimes causes longer-than-necessary delays.

All deposits within the range subsequently have their fill status
verified via the destination SpokePool contract, so the marginal cost
for including additional deposits is quite small.

Note also that deposit lookbacks are still enforced on the
SpokePoolClient search config, so looping mode is the only case where
this filtering would have been effective. The current direction of
looping mode is however to completely refresh the SpokePoolClient on
each update, so there shouldn't be any concerns about old deposits
appearing repeatedly.