refactor: Remove top-level bot loop (take two)
This change was previously merged and backed out shortly after because
it was seen to result in some bot instances hanging. This is almost
certainly because the original commit removed two instances of
process.exit(), and replaced them with prococess.exitCode = <X>. There
appear to be some instances where the bot can have some dangling
promises that take a long time to resolve, or that don't resolve for
some reason. Replaying the change, but without removing the
process.exit() calls should be OK.