feat(InventoryClient): Take repayment on chains that are over spoke target (#1411)
* feat(InventoryClient): Take repayment on chains that are over spoke target

This algorithm is based on discussions with Korpi.

The current algorithm for determining refund chain is:
1. if destination relayer balance is under target, take repayment on destination
2. Else if origin relayer balance is under target, take repayment on origin
3. Else mainnet

The proposed changes to the algorithm add two new rules above 1:
1. Sort all 7day chains (i.e. ORU's with 7 day withdrawal liveness) by the amount that their "latest running balances" are over their `spokePoolTarget` and evaluate them in this order. Remove any chains where the excess is < 0.
2. For each of these chains, if the relayer balance is under allocated on this chain, take repayment there
3. If there are no more 7 day chains, then fallback to existing algorithm where we evaluate destination chain and then origin chain relayer balance allocation

The complexity of this PR is computing "latest running balances". The algorithm I'm using is:
- Find running balance for token and chain using latest proposed root bundle containing token and chain
- Add all deposited `inputAmounts` to running balances that occurred after the latest proposal's end block
- Subtract all refunds `inputAmounts` from running balances that occurred after the latest proposal's end block

The last bullet point is obviously risky since it opens the relayer up to being griefed and tricked into taking repayment on the chain if someone sends an invalid fill with a large `inputAmount` on a chain with cheap gas. This is pretty easy to do though I'm not sure why someone one would do it. The fix is a change to the algorithm in `getApproximateRefundsForBlockRange` to match fills and deposits on `inputAmount` and `outputAmount` in addition to originChain and depositId. It still allows invalid fills to be sent but its a good approximation for now

* Update InventoryClient.ts

* refactor

* Update BundleDataClient.ts

* Add unit tests

* Switch signs

* Add unit test for getRunningBalanceExcesses

* Update InventoryClient.ts

* Update InventoryClient.ts

* Add 1.5x multipleir to target relayer allocation%

* Fix

* cache running balances, not pcts post relay

* Update InventoryClient.ts

* Update BundleDataClient.ts

* Update BundleDataClient.ts

* Update MockInventoryClient.ts

* Update InventoryClient.ts

* Update Relayer.ts

* Use loadData instead of loadDataFromArweave in getLatestProposedBundleData

* Add relayerTargetMulitplier to inventory config

* Refactor

* matt's comments

* Update InventoryClient.ts

* Update InventoryClient.ts

* default to destination chain if no token config for any eligible repayment chains

* Update InventoryClient.ts

* Don't get lp fee for chain we don't support in config

* Update Relayer.ts

* Update Relayer.ts

* Remove whitelisted relayers list param from getApproximateRefundsForBlockRange

There's no need to exit early now in the very fast getApproximateRefundsForBlockRange function

* don't compute running balances for chain-token combinations that aren't enabled

* refactor

* Update InventoryClient.ts

* Refactor SLOW_WITHDRAWAL_CHAINS

The inventory client should expose a function to get filtered list of slow withdrawal repayment chains

* Add getPossibleRepaymentChainIds to inventory client interface

This way relayer can call this function to better determine which chains it should compute LP fees for

* Fix tests

* Update InventoryClient.ts