feat(relayer): Externally-indexed SpokePoolClient (#1413)
This commit introduces a new relayer-specific SpokePoolClient that's
suited to running with a non-zero POLLING_DELAY. The benefits of this
change include:

- Significantly faster relayer loop times.
- Relayer performance is decoupled from RPC provider performance.
- Additional robustness to RPC provider errors & misbehaviour.
- Increased RPC speed & efficiency via WebSocket event subscription.
- Increased protection against re-orgs in looping mode.

In this mode, the relayer forks off one indexer process per chain. Each
indexer is responsible for indexing events based on a configurable
lookback, and subsequently subscribes to the required set of events
moving forward. Indexers supply event updates to the relayer after each
block update is emitted by the primary WebSocket provider; these events
are then retained in the SpokePoolClient to be used on the next
SpokePoolClient.update() call.

Because WebSocket providers are used for event subscription, the 
existing relayer ProviderUtils providers are not used. Instead, basic
quorum & finality checks are applied by the indexer before relaying the 
events to the relayer. There is still a possibility that reorgs occur
after events have been relayed to the relayer, so whenever a removal
notification is received by the indexer process it is immediately
relayed to the relayer so it can remove any stale events as soon as is
practical.

With this change it's possible to configure the relayer to be very
responsive to new deposit events. In testing it's been possible to make
fills reliably with only a minority of collisions. The sample size is
small due to a lack of test funds in the dev wallet, so deposit
confirmation tuning will be needed to hit the sweet spot of fast fills
that don't collide with the very-fast third-party relayers.

This logic can be enabled by adjusting the following config items:

Set:
- POLLING_DELAY to a positive number (i.e. POLLING_DELAY=10), and 
- RELAYER_EXTERNAL_SPOKEPOOL_INDEXER=true.