improve: Reduce mainnet bundle end block buffer to 5 blocks
We rarely if ever see re-orgs longer than 1 block, so 5 blocks is a reasonable value. This value controls how quickly after the last bundle the dataworker will propose the next one. The algorithm is here which waits to propose until the next proposed `mainnetBundleEndBlock > HEAD - bufferToPropose` where `bufferToPropose=5` in this PR.

Also, the worst thing that can happen with a re-org is that the proposal gets disputed because it includes a re-orged deposit