improve(relayer): Reduce delays & improve logging
This commit changes the way config.pollingDelay is used in the relayer.
Previously, the relayer would wait pollingDelay seconds _after_ its
previous run before commencing its next loop. This is OK when loop times
are fairly predictable, but in practice there can be quite high variance
in loop times due to variable RPC response timing. In cases where the
loop times are slow, the relayer becomes significantly less responsive
to new deposits.

Instead, use pollingDelay as an interval timer at which the relayer
should be started. In the event that the relayer takes more than
pollingDelay seconds to execute its loop, the delay will be skipped and
the relayer will instead proceed to evaluate and deposits that may have
occurred during the previous loop.

While here, improve the logging such that loop-related log messages are
only emitted in looping mode.