feat: Script to snapshot SpokePool EnabledDepositRoute events
This script is capable of scraping historical EnabledDepositRoute
events from all SpokePools configured via the ConfigStore. It has two
modes:

snapshot:
  Scrape all events and save them to a file.

verify:
  Read a snapshot from a file and verify it against the onchain state as
  at the point the snapshot was taken.

Example:
  yarn ts-node ./scripts/DepositRoutes.ts snapshot [--file <outputFile>]
  yarn ts-node ./scripts/DepositRoutes.ts verify [--file <outputFile>]

The HubPool chain ID can also optionally be supplied via the --chainId
option.

The intention of this script is to facilitate periodic snapshots to be
taken and stored in the relayer repository. The relayer would then load
these snapshots directly from disk, rather than querying back to
SpokePool genesis for each chain (which has ever-worse performance).
Each snapshot stores the blockNumber at which is was taken on each
chain, so the relayer is then able to request SpokePoolClient updates
from that point.

Scraping these events manually currently takes ~14 seconds in local
testing, even when the Redis cache has been primed with these events.
Loading from disk should take < 1 second, and is especially efficient
because only the events are retained in the file, whereas querying the
RPC provider requires the entire block range to be evaluated.

If this snapshotting process is done monthly or quarterly then it will
mitigate the incremental performance degradation that the relayer is
constantly subjected to.