improve: Allow BalanceAllocator users to force balance refresh
Users of the balance allocator (the Dataworker and the Monitor) use this client to track balance in order to decide when to proceed with an action. The issue is the rest of the bot's logic takes a while to run causing the balances fetched to become several minutes stale by the time the logic is acted upon.