fix(AcrossAPIClient): Reverse limits query direction
When querying limits, the API client current uses mainnet as the origin
chain. These limits are typically negative, implying that Across can
support _more_ volume than is available as liquid reserve in the
HubPool. This effectively renders the limits check redundant and could
cause the HubPool to become overcommitted.

While here, apply a slight refactor in recognition that API client
doesn't need require SpokePoolClient instances anymore, so drop them in
favour of a simple chain ID list. This removes the current constraint
that mainnet needs to be defined as a destination chain. This adjustment
makes the relayer more flexible and should be popular with
third-parties.