improve(Dataworker#index): Handle executor run not completing before subsequent proposal is sent
When the executor runs very slowly, say ~10 mins, then its possible that its run can start while a root bundle proposal is still pending, but in the middle of its run, the pending root bundle can be executed and the next bundle can be proposed.

The proposal would cause the current logic for detecting an `executorCollision` to give a false negative, resulting in an error suffered in the MulticallerClient.

This PR adds a check to `executorCollision` that the pending root bundle challenge timestamp is in the future in addition to its unclaimedLeafCount being positive.

An alternative implementation I considered was comparing the pending root bundle's `poolRebalanceRoot` with the one stored in the `HubPoolClient`. I think that would work as well.