improve(RelayerConfig): Permit symbol-based token config
This change permits the InventoryConfig to be specified in either
HubPool token address or token symbol. Token symbol is preferred because
it's simpler to configure and reduces the chance of mistakes.

During config, any symbols are resolved to their underlying token
addresses on the HubPool chain, so no subsequent changes are required
for any accesses into the post-processed inventory config.