feat(InventoryClient): Support 1:many HubPool mappings
This update to the InventoryClient permits multiple tokens to map to a
single HubPool token. This will be used with USDC/CCTP, where the
existing USDC deployment will map to both native USDC and bridged USDC
on the various non-mainnet chains supported by Across. See below for an
example of how this is configured. Note that the symbol _USDC is the
temporary label for native USDC.

  "USDC": {
    "USDC.e": {
      "10": { "targetPct": 8, "thresholdPct": 4 },
      "137": { "targetPct": 4, "thresholdPct": 1 },
      "324": { "targetPct": 8, "thresholdPct": 4 },
      "42161": { "targetPct": 8, "thresholdPct": 4 },
      "59144": { "targetPct": 5, "thresholdPct": 2 }
    },
    "USDbC": {
      "8453": { "targetPct": 5, "thresholdPct": 2 }
    },
    "_USDC": {
      "137": { "targetPct": 1, "thresholdPct": 1 }
    }
  },